Click and double click
Now to select an employee, you need to double click him.
To do a task is one click.

When the player double clicks it casts a raycast to the mouse position and if it detects the layer of the employee, it asigns that employee as the current employee; If it doesn't detect the employee layer it asigns the current employee as null.

When it is only one click, it cast a raycast that detects the layers of the flor and tells the employee to move towards the hit.point of the raycast.

Bugs: when double clicking, it also ejecutes the click event.