Changes in the UI
Added a UICamera and changed the canvas rendering mode to camera. This makes it more comfortable for us.
Replaced all UI text with TextMeshPro text
Changed the UI anchords.
Fixed an issue with orders: when orders were instantiated they didn't had a parent, that causes it to not spawn in the parent position. I fixed it by setting the parent in the instantiate method.
Changed the canvas screen match mode to match with screen size, the value is 0.5, so it scales equally.
Added a child called Gameplay in the canvas that is were all the gameplay ui will be.